feat: moved image_generation to services folder feat: renamed the auth_service to user_service and moved to services folder feat: restructured the architecture of the app
