+++ utils/common.py
@@ -1,4 +1,7 @@
 import bcrypt
+from PIL import Image
+import io
+import base64
 
 def hash_password(password):
     # Generate a salt
@@ -10,3 +13,12 @@ def hash_password(password):
 def check_password(password, hashed):
     # Check the password
     return bcrypt.checkpw(password.encode(), hashed)
+
+def pil_image_to_base64(image: Image.Image, format="JPEG") -> str:
+    if format not in ["JPEG", "PNG"]:
+        format = "JPEG"
+    image_stream = io.BytesIO()
+    image = image.convert("RGB")
+    image.save(image_stream, format=format)
+    base64_image = base64.b64encode(image_stream.getvalue()).decode("utf-8")
+    return base64_image
